+++ test/integration/test_add_peer.js
@@ -13,7 +13,7 @@ test(`Add a peer`, async ({end, equal}) => {
   const cluster = await createCluster({});
 
   const {lnd} = cluster.control;
-  const remoteNodeKey = cluster.remote_node_public_key;
+  const remoteNodeKey = cluster.remote.public_key;
 
   const connectedKeys = (await getPeers({lnd})).peers.map(n => n.public_key);